Create conda env for echolocatoR

Create a new env (or update and existing one) with the necessary Python, R, and command line packages to run echolocatoR.

create_echoR_env(
  conda_env = "echoR",
  python_version = NULL,
  channel = c("conda-forge", "bioconda", "r"),
  python_packages = c("pandas>=0.25.0", "pyarrow", "scikit-learn", "bitarray",
    "networkx", "rpy2", "scipy", "pandas-plink"),
  r_packages = c("r-base", "r-devtools"),
  cli_packages = c("tabix", "plink", "macs2"),
  force_install = FALSE,
  auth_token = devtools::github_pat()
)

Arguments

conda_env
The name of the conda env you want to create.

python_version
The version of Python to be used in this Conda environment. The associated Python package from Conda will be requested as python={python_version}. When NULL, the default python package will be used instead. For example, use python_version = "3.6" to request that the Conda environment be created with a copy of Python 3.6. This argument will be ignored if python is specified as part of the packages argument, for backwards compatibility.

channel
An optional character vector of Conda channels to include. When specified, the forge argument is ignored. If you need to specify multiple channels, including the Conda Forge, you can use c("conda-forge", <other channels>).

python_packages
Python packages to install.

r_packages
R packages to install.

cli_packages
Command Line Interface packages to install.

force_install
If env already exists, force create a new one.

auth_token
GitHub authentication token.

Details

By default uses echoR, the conda environment distributed with echolocatoR.

plinkhttps://anaconda.org/bioconda/plink

tabixhttps://anaconda.org/bioconda/tabix

See also