Install conda

Install conda if it has not already been installed.

install_conda(
  method = c("basilisk", "reticulate"),
  conda = "auto",
  verbose = TRUE,
  ...
)

Arguments

method

Method to use:

"basilisk"

"reticulate"

conda

The path to a conda executable. Use "auto" to allow reticulate to automatically find an appropriate conda binary. See Finding Conda and conda_binary() for more details.

verbose

Print messages.

...

Additional arguments to be passed to install_miniconda.

Examples

if (FALSE) { # \dontrun{
echoconda::install_conda()
} # }