Create conda env from yaml file

Create a conda environment from a yaml file. By default, creates the "echoR" conda env to support echolocatoR.

yaml_to_env(
  yaml_path = system.file(package = "echoconda", "conda", "echoR_mini.yml"),
  method = c("basilisk", "reticulate"),
  conda = "auto",
  force_new = FALSE,
  show_contents = FALSE,
  verbose = TRUE,
  ...
)

Source

GitHub Issue

StackOverflow suggestions for cross-platform troubleshooting

Arguments

yaml_path

Path to local or remote yaml file with conda build specifications.

method

Method to use to create the conda env:

"basilisk" :

Uses the R package basilisk.

"reticulate" :

Uses the R package reticulate.

"cli" :

Uses a custom wrapper for conda's command line interface.

conda

The path to a conda executable. Use "auto" to allow reticulate to automatically find an appropriate conda binary. See Finding Conda and conda_binary() for more details.

force_new

If the conda env already exists, overwrite it with a new one (DEFAULT: FALSE).

show_contents

Show the contents of the yaml file (if used).

verbose

Print messages.

...

Additional arguments passed to method-specific conda env build functions:

basilisk: create_env_basilisk

reticulate: conda_create

cli: create_env_cli

See also

Examples

if (FALSE) { # \dontrun{
conda_env <- yaml_to_env()
} # }