Create conda env from yaml file

Create a conda environment from a yaml file. By default, creates the "echoR" conda env to support echolocatoR.

yaml_to_env(
  yaml_path = system.file(package = "echoconda", "conda/echoR.yml"),
  conda = "auto",
  force_new = FALSE,
  verbose = TRUE
)

Source

GitHub Issue

StackOverflow suggestions for cross-platform troubleshooting

Arguments

yaml_path

Path to local or remote yaml file with conda build specifications.

conda

The path to a conda executable. Use "auto" to allow reticulate to automatically find an appropriate conda binary. See Finding Conda and conda_binary() for more details.

force_new

If the conda env already exists, overwrite it with a new one (DEFAULT: FALSE).

verbose

Print messages.

See also

Examples

conda_env <- echoconda::yaml_to_env()
#> echoconda:: Conda already installed.
#> echoconda:: Retrieving conda env name from yaml.
#> Identified yaml file stored in echoconda.
#> Yaml contents:
#> name: echoR
#> channels:
#>   - conda-forge
#>   - bioconda
#>   - nodefaults
#> dependencies:
#>   # Python
#>   - python>=3.6.1
#>   - pandas>=0.25.0
#>   - fastparquet
#>   - pyarrow
#>   - scipy
#>   - scikit-learn
#>   - tqdm
#>   - bitarray
#>   - networkx
#>   - rpy2
#>   - requests
#>   # Command line
#>   - htslib
#>   - plink
#>   - bcftools
#>   - wget
#>   - axel
#>   # R
#>   - r>=4.0.0
#>   - r-devtools
#>   - r-susier
#>   - r-reticulate
#>   - r-rlang
#>   - r-mvtnorm
#>   - r-desctools
#>   - r-exact
#>   - r-later
#>   - r-commonmark
#>   - r-patchwork
#>   - r-r.methodss3
#>   - r-brew
#>   - r-zip
#>   - r-xfun
#>   - r-credentials
#>   - r-whisker
#>   - r-gh
#>   - r-xopen
#>   - r-sessioninfo
#>   - r-promises
#>   - r-rex
#>   - r-crosstalk
#>   - r-roxygen2
#>   - r-gert
#>   - r-r.oo
#>   - r-rversions
#>   - r-rcmdcheck
#>   - r-dt
#>   - r-covr
#>   - r-usethis
#>   - r-openxlsx
#>   - r-r.utils
#>   - r-data.table
#>   - r-ggplot2
#>   - r-ckmeans.1d.dp
#>   - r-tidyverse
#>   - r-foreign
#>   - r-refgenome
#>   - r-xgr
#>   - r-coloc
#>   - r-susier
#>   ### There's also this implementation of susieR which includes later versions
#>   # - aksarkar::r-susier
#>   - r-biocmanager
#>   - bioconductor-snpstats
#>   - bioconductor-ggbio
#>   - bioconductor-ensdb.hsapiens.v75
#>   - bioconductor-biomart
#>   - radian
#>   - pip
#> 
#> echoconda:: Conda environment already exists: echoR
#> Time difference of 0.4 secs